feat(playbook): Removed redundancies and duplicated code
Removed all possible duplicated structs from playbook feature
Added necessary functionality from playbook feature structs to vcert structs
Moved relevant unit tests to vcert

Closes VC-24596
