+++ pkg/certificate/certificate.go
@@ -24,214 +24,24 @@ import (
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/x509"
-	"crypto/x509/pkix"
 	"encoding/pem"
 	"fmt"
-	"net"
-	"net/url"
-	"strings"
-	"time"
-
-	"github.com/Venafi/vcert/v4/pkg/util"
-	"github.com/youmark/pkcs8"
-
 	"reflect"
 	"sort"
+	"time"
 
+	"github.com/youmark/pkcs8"
+
+	"github.com/Venafi/vcert/v4/pkg/util"
 	"github.com/Venafi/vcert/v4/pkg/verror"
 )
 
-// EllipticCurve represents the types of supported elliptic curves
-type EllipticCurve int
-
-func (ec *EllipticCurve) String() string {
-	switch *ec {
-	case EllipticCurveP521:
-		return "P521"
-	case EllipticCurveP384:
-		return "P384"
-	case EllipticCurveP256:
-		return "P256"
-	case EllipticCurveED25519:
-		return "ED25519"
-	default:
-		return ""
-	}
-}
-
-// Set EllipticCurve value via a string
-func (ec *EllipticCurve) Set(value string) error {
-	switch strings.ToLower(value) {
-	case "p521", "p-521":
-		*ec = EllipticCurveP521
-	case "p384", "p-384":
-		*ec = EllipticCurveP384
-	case "p256", "p-256":
-		*ec = EllipticCurveP256
-	case "ed25519":
-		*ec = EllipticCurveED25519
-	default:
-		*ec = EllipticCurveDefault
-	}
-
-	return nil
-}
-
 const (
-	EllipticCurveNotSet EllipticCurve = iota
-	// EllipticCurveP521 represents the P521 curve
-	EllipticCurveP521
-	// EllipticCurveP256 represents the P256 curve
-	EllipticCurveP256
-	// EllipticCurveP384 represents the P384 curve
-	EllipticCurveP384
-	// EllipticED25519 represents the ED25519 curve
-	EllipticCurveED25519
-	EllipticCurveDefault = EllipticCurveP256
-
 	defaultRSAlength int = 2048
 )
 
-func AllSupportedCurves() []EllipticCurve {
-	return []EllipticCurve{EllipticCurveP521, EllipticCurveP256, EllipticCurveP384, EllipticCurveED25519}
-}
 func AllSupportedKeySizes() []int {
 	return []int{1024, 2048, 4096, 8192}
-}
-
-// KeyType represents the types of supported keys
-type KeyType int
-
-func (kt *KeyType) String() string {
-	switch *kt {
-	case KeyTypeRSA:
-		return "RSA"
-	case KeyTypeECDSA:
-		return "ECDSA"
-	case KeyTypeED25519:
-		return "ED25519"
-	default:
-		return ""
-	}
-}
-
-func (kt *KeyType) X509Type() x509.PublicKeyAlgorithm {
-	switch *kt {
-	case KeyTypeRSA:
-		return x509.RSA
-	case KeyTypeECDSA:
-		return x509.ECDSA
-	case KeyTypeED25519:
-		return x509.Ed25519
-	}
-	return x509.UnknownPublicKeyAlgorithm
-}
-
-// Set the key type via a string
-func (kt *KeyType) Set(value, curveValue string) error {
-	switch strings.ToLower(value) {
-	case "rsa":
-		*kt = KeyTypeRSA
-		return nil
-	case "ecdsa", "ec", "ecc":
-		curve := EllipticCurveNotSet
-		if err := curve.Set(curveValue); err != nil {
-			return err
-		}
-		if curve == EllipticCurveED25519 {
-			*kt = KeyTypeED25519
-			return nil
-		}
-
-		*kt = KeyTypeECDSA
-		return nil
-	}
-	return fmt.Errorf("%w: unknown key type: %s", verror.VcertError, value) //todo: check all calls
-}
-
-const (
-	// KeyTypeRSA represents a key type of RSA
-	KeyTypeRSA KeyType = iota
-	// KeyTypeECDSA represents a key type of ECDSA
-	KeyTypeECDSA
-	// KeyTypeED25519 represents a key type of ED25519
-	KeyTypeED25519
-)
-
-type CSrOriginOption int
-
-const (
-	// LocalGeneratedCSR - this vcert library generates CSR internally based on Request data
-	LocalGeneratedCSR CSrOriginOption = iota // local generation is default.
-	// ServiceGeneratedCSR - server generate CSR internally based on zone configuration and data from Request
-	ServiceGeneratedCSR
-	// UserProvidedCSR - client provides CSR from external resource and vcert library just check and send this CSR to server
-	UserProvidedCSR
-)
-
-type CustomFieldType int
-
-const (
-	CustomFieldPlain CustomFieldType = 0 + iota
-	CustomFieldOrigin
-)
-
-// CustomField can be used for adding additional information to certificate. For example: custom fields or Origin.
-// By default it's custom field. For adding Origin set Type: CustomFieldOrigin
-// For adding custom field with one name and few values give to request:
-//
-//	request.CustomFields = []CustomField{
-//	  {Name: "name1", Value: "value1"}
-//	  {Name: "name1", Value: "value2"}
-//	}
-type CustomField struct {
-	Type  CustomFieldType
-	Name  string
-	Value string
-}
-
-type Location struct {
-	Instance, Workload, TLSAddress string
-	Replace                        bool
-}
-
-// Request contains data needed to generate a certificate request
-// CSR is a PEM-encoded Certificate Signing Request
-type Request struct {
-	CADN               string
-	Subject            pkix.Name
-	DNSNames           []string
-	OmitSANs           bool
-	EmailAddresses     []string
-	IPAddresses        []net.IP
-	URIs               []*url.URL
-	UPNs               []string
-	Attributes         []pkix.AttributeTypeAndValueSET
-	SignatureAlgorithm x509.SignatureAlgorithm
-	FriendlyName       string
-	KeyType            KeyType
-	KeyLength          int
-	KeyCurve           EllipticCurve
-	csr                []byte // should be a PEM-encoded CSR
-	PrivateKey         crypto.Signer
-	CsrOrigin          CSrOriginOption
-	PickupID           string
-	//Cloud Certificate ID
-	CertID          string
-	ChainOption     ChainOption
-	KeyPassword     string
-	FetchPrivateKey bool
-	/*	Thumbprint is here because *Request is used in RetrieveCertificate().
-		Code should be refactored so that RetrieveCertificate() uses some abstract search object, instead of *Request{PickupID} */
-	Thumbprint       string
-	Timeout          time.Duration
-	CustomFields     []CustomField
-	Location         *Location
-	ValidityDuration *time.Duration
-	IssuerHint       util.IssuerHint
-
-	// DEPRECATED: use ValidityDuration instead, this field is ignored if ValidityDuration is set
-	ValidityHours int
 }
 
 //SSH Certificate structures
@@ -477,35 +287,9 @@ type CertSeachInfo struct {
 	CertificateRequestGuid string `json:"Guid"`
 }
 
-// SetCSR sets CSR from PEM or DER format
-func (request *Request) SetCSR(csr []byte) error {
-	pemBlock, _ := pem.Decode(csr)
-	if pemBlock != nil {
-		if strings.HasSuffix(pemBlock.Type, "CERTIFICATE REQUEST") {
-			request.csr = csr
-			return nil
-		}
-	}
-
-	//Determine CSR type and use appropriate function
-	parsedCSR, err := x509.ParseCertificateRequest(csr)
-	if err != nil {
-		return err
-	}
-	if parsedCSR != nil {
-		request.csr = pem.EncodeToMemory(GetCertificateRequestPEMBlock(csr))
-		return nil
-	}
-	return fmt.Errorf("%w: can't determine CSR type for %s", verror.UserDataError, csr)
-}
-
-// GetCSR returns CSR in PEM format
-func (request Request) GetCSR() []byte {
-	return request.csr
-}
-
-// GenerateRequest generates a certificate request
+// Deprecated: GenerateRequest is deprecated
 // Please use method Request.GenerateCSR()
+// GenerateRequest generates a certificate request
 // TODO: Remove usage from all libraries, deprecated
 func GenerateRequest(request *Request, privateKey crypto.Signer) error {
 	pk := request.PrivateKey
@@ -513,128 +297,6 @@ func GenerateRequest(request *Request, privateKey crypto.Signer) error {
 	err := request.GenerateCSR()
 	request.PrivateKey = pk
 	return err
-}
-
-// GenerateCSR creates CSR for sending to server based on data from Request fields. It rewrites CSR field if it`s already filled.
-func (request *Request) GenerateCSR() error {
-	certificateRequest := x509.CertificateRequest{}
-	certificateRequest.Subject = request.Subject
-	if !request.OmitSANs {
-		addSubjectAltNames(&certificateRequest, request.DNSNames, request.EmailAddresses, request.IPAddresses, request.URIs, request.UPNs)
-	}
-	certificateRequest.Attributes = request.Attributes
-
-	csr, err := x509.CreateCertificateRequest(rand.Reader, &certificateRequest, request.PrivateKey)
-	if err != nil {
-		csr = nil
-	}
-	err = request.SetCSR(csr)
-	//request.CSR = pem.EncodeToMemory(GetCertificateRequestPEMBlock(csr))
-	return err
-}
-
-// GeneratePrivateKey creates private key (if it doesn`t already exist) based on request.KeyType, request.KeyLength and request.KeyCurve fileds
-func (request *Request) GeneratePrivateKey() error {
-	if request.PrivateKey != nil {
-		return nil
-	}
-	var err error
-	switch request.KeyType {
-	case KeyTypeECDSA:
-		request.PrivateKey, err = GenerateECDSAPrivateKey(request.KeyCurve)
-	case KeyTypeED25519:
-		request.PrivateKey, err = GenerateED25519PrivateKey()
-	case KeyTypeRSA:
-		if request.KeyLength == 0 {
-			request.KeyLength = defaultRSAlength
-		}
-		if request.KeyLength < AllSupportedKeySizes()[0] {
-			return fmt.Errorf("key Size must be %d or greater. But it is %d", AllSupportedKeySizes()[0], request.KeyLength)
-		}
-		request.PrivateKey, err = GenerateRSAPrivateKey(request.KeyLength)
-	default:
-		return fmt.Errorf("%w: unable to generate certificate request, key type %s is not supported", verror.VcertError, request.KeyType.String())
-	}
-	return err
-}
-
-// CheckCertificate validate that certificate returned by server matches data in request object. It can be used for control server.
-func (request *Request) CheckCertificate(certPEM string) error {
-	pemBlock, _ := pem.Decode([]byte(certPEM))
-	if pemBlock == nil {
-		return fmt.Errorf("%w: invalid pem format certificate %s", verror.CertificateCheckError, certPEM)
-	}
-	if pemBlock.Type != "CERTIFICATE" {
-		return fmt.Errorf("%w: invalid pem type %s (expect CERTIFICATE)", verror.CertificateCheckError, pemBlock.Type)
-	}
-	cert, err := x509.ParseCertificate(pemBlock.Bytes)
-	if err != nil {
-		return err
-	}
-	if request.PrivateKey != nil {
-		if request.KeyType.X509Type() != cert.PublicKeyAlgorithm {
-			return fmt.Errorf("%w: unmatched key type: %s, %s", verror.CertificateCheckError, request.KeyType.X509Type(), cert.PublicKeyAlgorithm)
-		}
-		switch cert.PublicKeyAlgorithm {
-		case x509.RSA:
-			certPubKey := cert.PublicKey.(*rsa.PublicKey)
-			reqPubkey, ok := request.PrivateKey.Public().(*rsa.PublicKey)
-			if !ok {
-				return fmt.Errorf("%w: request KeyType not matched with real PrivateKey type", verror.CertificateCheckError)
-			}
-
-			if certPubKey.N.Cmp(reqPubkey.N) != 0 {
-				return fmt.Errorf("%w: unmatched key modulus", verror.CertificateCheckError)
-			}
-		case x509.ECDSA:
-			certPubkey := cert.PublicKey.(*ecdsa.PublicKey)
-			reqPubkey, ok := request.PrivateKey.Public().(*ecdsa.PublicKey)
-			if !ok {
-				return fmt.Errorf("%w: request KeyType not matched with real PrivateKey type", verror.CertificateCheckError)
-			}
-			if certPubkey.X.Cmp(reqPubkey.X) != 0 {
-				return fmt.Errorf("%w: unmatched X for elliptic keys", verror.CertificateCheckError)
-			}
-		case x509.Ed25519:
-			certPubkey := cert.PublicKey.(ed25519.PublicKey)
-			reqPubkey, ok := request.PrivateKey.Public().(ed25519.PublicKey)
-			if !ok {
-				return fmt.Errorf("%w: request KeyType not matched with real PrivateKey type", verror.CertificateCheckError)
-			}
-			if !certPubkey.Equal(reqPubkey) {
-				return fmt.Errorf("%w: unmatched elliptic ed25519 keys", verror.CertificateCheckError)
-			}
-		default:
-			return fmt.Errorf("%w: unknown key algorythm %d", verror.CertificateCheckError, cert.PublicKeyAlgorithm)
-		}
-	} else if len(request.csr) != 0 {
-		pemBlock, _ := pem.Decode(request.csr)
-		if pemBlock == nil {
-			return fmt.Errorf("%w: bad CSR: %s", verror.CertificateCheckError, string(request.csr))
-		}
-		csr, err := x509.ParseCertificateRequest(pemBlock.Bytes)
-		if err != nil {
-			return err
-		}
-		if cert.PublicKeyAlgorithm != csr.PublicKeyAlgorithm {
-			return fmt.Errorf("%w: unmatched key type: %s, %s", verror.CertificateCheckError, cert.PublicKeyAlgorithm, csr.PublicKeyAlgorithm)
-		}
-		switch csr.PublicKeyAlgorithm {
-		case x509.RSA:
-			certPubKey := cert.PublicKey.(*rsa.PublicKey)
-			reqPubKey := csr.PublicKey.(*rsa.PublicKey)
-			if certPubKey.N.Cmp(reqPubKey.N) != 0 {
-				return fmt.Errorf("%w: unmatched key modulus", verror.CertificateCheckError)
-			}
-		case x509.ECDSA:
-			certPubKey := cert.PublicKey.(*ecdsa.PublicKey)
-			reqPubKey := csr.PublicKey.(*ecdsa.PublicKey)
-			if certPubKey.X.Cmp(reqPubKey.X) != 0 {
-				return fmt.Errorf("%w: unmatched X for elliptic keys", verror.CertificateCheckError)
-			}
-		}
-	}
-	return nil
 }
 
 func publicKey(priv crypto.Signer) crypto.PublicKey {
@@ -825,7 +487,7 @@ func NewRequest(cert *x509.Certificate) *Request {
 	return req
 }
 
-// find a certificate from a list of certificates whose Sans.DNS matches and is
+// FindNewestCertificateWithSans finds a certificate from a list of certificates whose Sans.DNS matches and is
 // the newest
 func FindNewestCertificateWithSans(certificates []*CertificateInfo, sans_ *Sans) (*CertificateInfo, error) {
 	sans := Sans{}
