+++ pkg/certificate/chainOption_test.go
@@ -1,10 +1,9 @@
-package certrequest
+package certificate
 
 import (
 	"fmt"
 	"testing"
 
-	vcert "github.com/Venafi/vcert/v4/pkg/certificate"
 	"github.com/stretchr/testify/suite"
 	"gopkg.in/yaml.v3"
 )
@@ -15,7 +14,6 @@ type ChainOptionSuite struct {
 	testCases []struct {
 		chainOption ChainOption
 		strValue    string
-		vcertValue  vcert.ChainOption
 	}
 }
 
@@ -23,12 +21,10 @@ func (s *ChainOptionSuite) SetupTest() {
 	s.testCases = []struct {
 		chainOption ChainOption
 		strValue    string
-		vcertValue  vcert.ChainOption
 	}{
-		{chainOption: ChainOptionIgnore, strValue: strChainOptionIgnore, vcertValue: vcert.ChainOptionIgnore},
-		{chainOption: ChainOptionRootFirst, strValue: strChainOptionRootFirst, vcertValue: vcert.ChainOptionRootFirst},
-		{chainOption: ChainOptionRootLast, strValue: strChainOptionRootLast, vcertValue: vcert.ChainOptionRootLast},
-		{chainOption: ChainOptionUnknown, strValue: strChainOptionUnknown, vcertValue: vcert.ChainOptionRootLast},
+		{chainOption: ChainOptionIgnore, strValue: strChainOptionIgnore},
+		{chainOption: ChainOptionRootFirst, strValue: strChainOptionRootFirst},
+		{chainOption: ChainOptionRootLast, strValue: strChainOptionRootLast},
 	}
 
 	s.testYaml = `---
@@ -60,24 +56,19 @@ func (s *ChainOptionSuite) TestChainOption_String() {
 	}
 }
 
-func (s *ChainOptionSuite) TestChainOption_ToVCert() {
-	for _, tc := range s.testCases {
-		s.Run(tc.strValue, func() {
-			vcertCo := tc.chainOption.ToVCert()
-			s.Equal(tc.vcertValue, vcertCo)
-		})
-	}
-}
-
 func (s *ChainOptionSuite) TestChainOption_UnmarshalYAML() {
 	for _, tc := range s.testCases {
 		s.Run(tc.strValue, func() {
-			var r Request
+			result := struct {
+				Cn          string      `yaml:"cn"`
+				ChainOption ChainOption `yaml:"chainOption"`
+			}{}
+
 			parsedYaml := fmt.Sprintf(s.testYaml, tc.strValue)
-			err := yaml.Unmarshal([]byte(parsedYaml), &r)
+			err := yaml.Unmarshal([]byte(parsedYaml), &result)
 
 			s.Nil(err)
-			s.Equal(tc.chainOption, r.ChainOption)
+			s.Equal(tc.chainOption, result.ChainOption)
 		})
 	}
 }
