+++ pkg/playbook/app/installer/capi.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/Venafi/vcert/v4/pkg/certificate"
 	"github.com/Venafi/vcert/v4/pkg/playbook/app/domain"
-	"github.com/Venafi/vcert/v4/pkg/playbook/app/domain/certrequest"
 	"github.com/Venafi/vcert/v4/pkg/playbook/util"
 )
 
@@ -29,7 +28,7 @@ func NewCAPIInstaller(inst domain.Installation) CAPIInstaller {
 // 1. Does the certificate exists? > Install if it doesn't.
 // 2. Does the certificate is about to expire? Renew if about to expire.
 // Returns true if the certificate needs to be installed.
-func (r CAPIInstaller) Check(_ string, renewBefore string, request certrequest.Request) (bool, error) {
+func (r CAPIInstaller) Check(_ string, renewBefore string, request domain.PlaybookRequest) (bool, error) {
 	zap.L().Debug(fmt.Sprintf("checking certificate at: %s", r.Location))
 
 	friendlyName := request.Subject.CommonName
