+++ pkg/playbook/app/installer/pem.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/Venafi/vcert/v4/pkg/certificate"
 	"github.com/Venafi/vcert/v4/pkg/playbook/app/domain"
-	"github.com/Venafi/vcert/v4/pkg/playbook/app/domain/certrequest"
 	"github.com/Venafi/vcert/v4/pkg/playbook/util"
 )
 
@@ -31,7 +30,7 @@ func NewPEMInstaller(inst domain.Installation) PEMInstaller {
 // 1. Does the certificate exists? > Install if it doesn't.
 // 2. Does the certificate is about to expire? Renew if about to expire.
 // Returns true if the certificate needs to be installed.
-func (r PEMInstaller) Check(_ string, renewBefore string, _ certrequest.Request) (bool, error) {
+func (r PEMInstaller) Check(_ string, renewBefore string, _ domain.PlaybookRequest) (bool, error) {
 	certPath := filepath.Join(r.Location, r.PEMCertFilename)
 	zap.L().Debug(fmt.Sprintf("checking certificate at: %s", certPath))
