+++ pkg/playbook/app/installer/pkcs12.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/Venafi/vcert/v4/pkg/certificate"
 	"github.com/Venafi/vcert/v4/pkg/playbook/app/domain"
-	"github.com/Venafi/vcert/v4/pkg/playbook/app/domain/certrequest"
 	"github.com/Venafi/vcert/v4/pkg/playbook/util"
 )
 
@@ -30,7 +29,7 @@ func NewPKCS12Installer(inst domain.Installation) PKCS12Installer {
 // 1. Does the certificate exists? > Install if it doesn't.
 // 2. Does the certificate is about to expire? Renew if about to expire.
 // Returns true if the certificate needs to be installed.
-func (r PKCS12Installer) Check(_ string, renewBefore string, request certrequest.Request) (bool, error) {
+func (r PKCS12Installer) Check(_ string, renewBefore string, request domain.PlaybookRequest) (bool, error) {
 	zap.L().Debug(fmt.Sprintf("checking certificate at: %s", r.Location))
 
 	// Check certificate file exists
